Split the candidate's return half out of its establishment half
`workspace_candidate.py` owned both what a candidate workspace is and
what becomes of it. `integrate` and `retire` are the second thing: they
read and write no ticket, they run at landing rather than at launch, and
their ordering constraint is with each other, not with `establish`. They
move to `scripts/workspace_return.py` unchanged.

`_git_out` and `_branch_tip` are read by both halves, so they move down
to their natural owner, `workspace_git.py`, rather than being copied.
Mechanical: no behavior changes, and the four trunk defects the next
commits fix are all in the half that just moved.
